Fix pre-commit issues
- Fix import order in bash.py
- Make cwd property abstract in base class to match implementations
- Address trailing whitespace and formatting issues
